// gtfs-processor/shared/gtfs-db-types/types.ts
export enum Direction {
  OUTBOUND = "OUTBOUND",
  INBOUND = "INBOUND",
}
// ...
export enum PickupDropOff {
  REGULAR = "REGULAR",
  NO_PICKUP_OR_DROP_OFF = "NO_PICKUP_OR_DROP_OFF",
  PHONE_AGENCY = "PHONE_AGENCY",
  COORDINATE_WITH_DRIVER = "COORDINATE_WITH_DRIVER",
}

export enum RouteType {
  TRAM = "TRAM",
  SUBWAY = "SUBWAY",
  RAIL = "RAIL",
  BUS = "BUS",
  FERRY = "FERRY",
  CABLE_TRAM = "CABLE_TRAM",
  AERIAL_LIFT = "AERIAL_LIFT",
  FUNICULAR = "FUNICULAR",
  TROLLEYBUS = "TROLLEYBUS",
  MONORAIL = "MONORAIL",
}

export enum StopTimeUpdateScheduleRelationship {
  SCHEDULED = "SCHEDULED",
  SKIPPED = "SKIPPED",
  NO_DATA = "NO_DATA",
  UNSCHEDULED = "UNSCHEDULED",
}

export enum StrEnum {

}

export enum Timepoint {
  APPROXIMATE = "APPROXIMATE",
  EXACT = "EXACT",
}

export enum TripDescriptorScheduleRelationship {
  SCHEDULED = "SCHEDULED",
  ADDED = "ADDED",
  UNSCHEDULED = "UNSCHEDULED",
  CANCELED = "CANCELED",
  REPLACEMENT = "REPLACEMENT",
  DUPLICATED = "DUPLICATED",
  NEW = "NEW",
  DELETED = "DELETED",
}

export enum TripInstanceState {
  PRISTINE = "PRISTINE",
  DIRTY = "DIRTY",
  REMOVED = "REMOVED",
}
// ...
export interface StopTimeInstance {
  stop_id: string;
  stop_headsign?: null | string;
  pickup_type?: PickupDropOff | null;
  drop_off_type?: PickupDropOff | null;
  timepoint: Timepoint;
  shape_dist_traveled?: null | number;
  arrival_datetime?: Date | null;
  departure_datetime?: Date | null;
  predicted_arrival_datetime?: Date | null;
  predicted_departure_datetime?: Date | null;
  predicted_arrival_uncertainty?: null | number;
  predicted_departure_uncertainty?: null | number;
  schedule_relationship?: StopTimeUpdateScheduleRelationship | null;
}

export interface Trip {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  trip_id: string;
  route_id: string;
  service_id: string;
  trip_headsign?: null | string;
  trip_short_name?: null | string;
  direction_id?: Direction | null;
  block_id?: null | string;
  shape_id?: null | string;
}

export interface TripDescriptor {
  trip_id?: null | string;
  start_time?: null | string;
  start_date?: null | string;
  route_id?: null | string;
  direction_id?: Direction | null;
  schedule_relationship?: TripDescriptorScheduleRelationship | null;
}

export interface TripInstance {
  _id?: null | string;
  revision_id?: null | string;
  agency_id: string;
  trip_id: string;
  start_date: string;
  start_time: string;
  route_id: string;
  direction_id?: Direction | null;
  state?: TripInstanceState;
  start_datetime: Date;
  stop_times: StopTimeInstance[];
  trip: any;
  route: any;
  shape?: any | null;
  vehicle?: any | null;
  positions?: any[];
}